correction erreur de débutantgit add .! nulle nulle

+++ src/main/java/com/berta/magasin/repositories/IClientRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.berta.magasin.models.Client;
 
-public interface IClientRepository extends JpaRepository<Client, Long>{
+public interface IClientRepository extends JpaRepository<Client, Integer>{
 
 }
